Enemy/bossManta: body message/battle state machines (0.9% -> 96.9%, 0.5% -> real logic)
TMantaMessageState::update: 0.9% -> 96.9%. Spawn-demo balloon prompt
(nerve check via TSpineBase::getLatestNerve), alive-enemy-count
threshold balloon, and battle-phase balloon, each gated on unk4 and
firing gpMarDirector->getConsole()->startAppearBalloon. Residual is a
switch-dispatch pivot-order difference MWCC makes for the dense
0/1/2 case set (same class of do-last codegen choice as elsewhere).

TMantaBattleState::update: replaces the no-op stub with real logic
for all 4 states - appear-demo camera trigger + first-generation
spawn, generation-5 escalation once all mantas reach max generation,
victory SE once all gen-5 mantas are dead, and a post-victory guard
on the SE handle. Last case's final event-flag lookup (searching a
named object and setting a byte field) is left TODO - the exact
type/field semantics weren't confirmed with enough certainty to guess
without risking a wrong match; everything else is real, non-stub
logic. Retyped TMantaMessageState::unk4 to s32 (word-load in target).

+++ include/Enemy/BossManta.hpp
@@ -108,7 +108,7 @@ public:
 	public:
 		void update();
 		TBossMantaManager* unk0;
-		u8 unk4;
+		s32 unk4;
 	};
 	class TMantaBattleState {
 	public:
